reorganizae tooltip and data display
