+++ containers/ecr-viewer/src/app/services/evaluateFhirDataService.ts
@@ -16,7 +16,7 @@ import {
 } from "./formatService";
 import fhirpath_r4_model from "fhirpath/fhir-context/r4";
 import { Element } from "fhir/r4";
-import { DisplayDataProps } from "@/app/DataDisplay";
+import { DisplayDataProps } from "@/app/view-data/components/DataDisplay";
 
 /**
  * Evaluates patient name from the FHIR bundle and formats it into structured data for display.
